Added `de/options` module
